Patch for DummyFiles bug
There was an issue with dummy files not being found when trying to
upload a project.  This could have been caused by a previous crash not
creating the necessary dummy files.  This will need to be addresssed in
the next update.

+++ src/DummyFiles.py
@@ -78,7 +78,7 @@ class DummyFiles:
 
         for file in json_file['dummy']:
             file = Path(f"{self.filepath}/{file}")
-            if file.stat().st_size == 0:
+            if file.exists() and file.stat().st_size == 0:
                 file.unlink()
 
         json_file['dummy'] = []
